robotic_control/requirements.txt
torch==2.1.0
numpy==1.26.1
gymnasium==0.29.1
pybullet==3.2.5
rl-zoo3==2.1.0
stable-baselines3==2.1.0
shimmy==1.3.0
torchmetrics
matplotlib